Split index.js into small modules
